+++ src/main/java/gift/member/dto/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package gift.user.dto;
+package gift.member.dto;
 
 public record LoginResponseDto(
     String token
