+++ src/main/java/gift/member/dto/RegisterRequestDto.java
@@ -1,4 +1,4 @@
-package gift.user.dto;
+package gift.member.dto;
 
 import jakarta.validation.constraints.NotBlank;
