+++ src/main/java/gift/member/dto/RegisterResponseDto.java
@@ -1,4 +1,4 @@
-package gift.user.dto;
+package gift.member.dto;
 
 public record RegisterResponseDto(
     String token
