+++ src/main/java/gift/member/exception/MemberNotFoundException.java
@@ -1,0 +1,8 @@
+package gift.member.exception;
+
+public class MemberNotFoundException extends RuntimeException {
+
+    public MemberNotFoundException(String message) {
+        super(message);
+    }
+}
