+++ src/main/java/gift/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package gift.member.repository;
+
+import gift.member.entity.Member;
+
+public interface MemberRepository {
+
+    void saveMember(Member member);
+
+    Member findMemberByEmail(String email);
+}
