+++ src/main/java/gift/member/service/MemberService.java
@@ -1,0 +1,13 @@
+package gift.member.service;
+
+import gift.member.dto.LoginRequestDto;
+import gift.member.dto.LoginResponseDto;
+import gift.member.dto.RegisterRequestDto;
+import gift.member.dto.RegisterResponseDto;
+
+public interface MemberService {
+
+    RegisterResponseDto registerMember(RegisterRequestDto registerRequestDto);
+
+    LoginResponseDto login(LoginRequestDto loginRequestDto);
+}
