--- src/main/java/gift/user/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package gift.user.repository;
-
-import gift.user.entity.Member;
-
-public interface MemberRepository {
-
-    void saveMember(Member member);
-
-}
