--- src/main/java/gift/user/service/MemberService.java
@@ -1,13 +1,0 @@
-package gift.user.service;
-
-import gift.user.dto.LoginRequestDto;
-import gift.user.dto.LoginResponseDto;
-import gift.user.dto.RegisterRequestDto;
-import gift.user.dto.RegisterResponseDto;
-
-public interface MemberService {
-
-    RegisterResponseDto registerMember(RegisterRequestDto registerRequestDto);
-
-    LoginResponseDto login(LoginRequestDto loginRequestDto);
-}
